reconfig fp16 in adaconv_patch

+++ net.py
@@ -924,7 +924,7 @@ def content_feature_contrastive(input, decoder):
 def compute_contrastive_loss(feat_q, feat_k, tau, index):
     out = torch.mm(feat_q, feat_k.transpose(1, 0)) / tau
     #loss = self.cross_entropy_loss(out, torch.zeros(out.size(0), dtype=torch.long, device=feat_q.device))
-    loss = F.binary_cross_entropy_with_logits(out, torch.tensor([index], dtype=torch.long, device=feat_q.device).expand(1,4))
+    loss = F.binary_cross_entropy_with_logits(out, torch.tensor([index], device=feat_q.device).expand(1,4))
     return loss
